# SNOW-1926348: Re-implement groupby-apply without pivot