<!---
Please answer these questions before creating your pull request. Thanks!
--->

1. Which Jira issue is this PR addressing? Make sure that there is an accompanying issue to your PR.

   <!---
   In this section, please add a Snowflake Jira issue number.

   Note that if a corresponding GitHub issue exists, you should still include
   the Snowflake Jira issue number. For example, for GitHub issue
   https://github.com/snowflakedb/snowpark-python/issues/1400, you should
   add "SNOW-1335071" here.
    --->

   Fixes SNOW-1926348

2. Fill out the following pre-review checklist:

   - [ ] I am adding a new automated test(s) to verify correctness of my new code
      - [ ] If this test skips Local Testing mode, I'm requesting review from @snowflakedb/local-testing
   - [ ] I am adding new logging messages
   - [ ] I am adding a new telemetry message
   - [ ] I am adding new credentials
   - [ ] I am adding a new dependency
   - [ ] If this is a new feature/behavior, I'm adding the Local Testing parity changes.
   - [ ] I acknowledge that I have ensured my changes to be thread-safe. Follow the link for more information: [Thread-safe Developer Guidelines](https://github.com/snowflakedb/snowpark-python/blob/main/CONTRIBUTING.md#thread-safe-development)
   - [ ] If adding any arguments to public Snowpark APIs or creating new public Snowpark APIs, I acknowledge that I have ensured my changes include AST support. Follow the link for more information: [AST Support Guidelines](https://github.com/snowflakedb/snowpark-python/blob/main/CONTRIBUTING.md#ast-abstract-syntax-tree-support-in-snowpark)

3. Please describe how your code solves the related issue.

 This PR re-implements the `groupby-apply` method without the use of pivot. We infer the output schema and use that to create the UDTF with right number of columns. If we fail to infer to schema we fallback to old implementation.
 
**Performance improvement numbers:**
 microbenchmark dataframe_groupby_apply_transform Before: 64.9 sec After: 21.4 sec
 microbenchmark dataframe_groupby_apply_concat Before: 106.2 sec After: 40.7 sec
 microbenchmark dataframe_groupby_apply_agg Before: 14.0 sec After: 13.4 sec
 
 Customer support ticket [Believe](https://snowflakecomputing.atlassian.net/browse/SNOW-1947974):
 [Before](https://app.snowflake.com/sfengineering/sfctest0_aws_us_west_2/#/compute/history/queries?user=NAMED%3ASNOWPANDAS_TEST&start=1743145200000&columns=sqlText%2CqueryId%2Cstatus%2Cuser%2Cwarehouse%2CtotalDuration%2CstartTime&preset=PRESET_LAST_3_DAYS&type=relative&relative=%7B%22tense%22%3A%22past%22%2C%22value%22%3A3%2C%22unit%22%3A%22day%22%2C%22excludePartial%22%3Afalse%2C%22exclusionSize%22%3A%22day%22%2C%22exclusionSizeParam%22%3A%22%22%7D&startDate=2025-03-30T00%3A00%3A00Z&endDate=2025-03-31T00%3A00%3A00Z&page=history%2Fqueries&session_id=3563545552846631):  6 min 21 secs
 [After](https://app.snowflake.com/sfengineering/sfctest0_aws_us_west_2/#/compute/history/queries?user=NAMED%3ASNOWPANDAS_TEST&start=1743145200000&columns=sqlText%2CqueryId%2Cstatus%2Cuser%2Cwarehouse%2CtotalDuration%2CstartTime&preset=PRESET_LAST_3_DAYS&type=relative&relative=%7B%22tense%22%3A%22past%22%2C%22value%22%3A3%2C%22unit%22%3A%22day%22%2C%22excludePartial%22%3Afalse%2C%22exclusionSize%22%3A%22day%22%2C%22exclusionSizeParam%22%3A%22%22%7D&startDate=2025-03-30T00%3A00%3A00Z&endDate=2025-03-31T00%3A00%3A00Z&page=history%2Fqueries&session_id=3563545552865851): 2 min 19 secs
 
 To reduce the scope following will be done in separate PRs:
 we don't infer the columns types yet, that will be a follow up PR.
 We still use old implementation for DataFrame.apply. In a follow up PR I will use new implementation for this as well.