@sfc-gh-snowflakedb-snyk-sa 2025-03-29
![](https://res.cloudinary.com/snyk/image/upload/r-d/scm-platform/snyk-pull-requests/pr-banner-default.svg)

### :tada: **Snyk checks have passed. No issues have been found so far.**

:white_check_mark: **security/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/accf72dd-70e7-4adf-aaee-8e79e94b3571?refs=77fc2c653c1edfbac00e14776a12e0e8601411a5&source=prChecksComment)

:white_check_mark: **license/snyk** check is complete. No issues have been found. [(View Details)](https://app.snyk.io/org/snowflakedb-sca-scanning-public-repo/pr-checks/accf72dd-70e7-4adf-aaee-8e79e94b3571/license?refs=77fc2c653c1edfbac00e14776a12e0e8601411a5&source=prChecksComment)